Crate colour

Macros for creating coloured console output.

The macros are based on the following colours:

grey

red

green

yellow

blue

magenta

cyan

black

white

There are dark versions of each colour except black and white, and for all colours there is a version with suffix _ln which appends a newline (similar to print! and println!).

There are also prnt! and prnt_ln! available which print using the current default foreground colour.

Every variant also has a version with prefix e_. Variants with this prefix output to stderr, while those without the e_ prefix output to stdout.

Macros

black

black_ln

blue

blue_ln

cyan

cyan_ln

dark_blue

dark_blue_ln

dark_cyan

dark_cyan_ln

dark_green

dark_green_ln

dark_grey

dark_grey_ln

dark_magenta

dark_magenta_ln

dark_red

dark_red_ln

dark_yellow

dark_yellow_ln

e_black

e_black_ln

e_blue

e_blue_ln

e_cyan

e_cyan_ln

e_dark_blue

e_dark_blue_ln

e_dark_cyan

e_dark_cyan_ln

e_dark_green

e_dark_green_ln

e_dark_grey

e_dark_grey_ln

e_dark_magenta

e_dark_magenta_ln

e_dark_red

e_dark_red_ln

e_dark_yellow

e_dark_yellow_ln

e_green

e_green_ln

e_grey

e_grey_ln

e_magenta

e_magenta_ln

e_prnt

e_prnt_ln

e_red

e_red_ln

e_white

e_white_ln

e_yellow

e_yellow_ln

green

green_ln

grey

grey_ln

magenta

magenta_ln

prnt

prnt_ln

red

red_ln

white

white_ln

yellow

yellow_ln