Crate colour

Macros for creating coloured console output.

The macros are based on the following colours:

grey

red

green

yellow

blue

magenta

cyan

black

white

There are dark versions of each colour, and for all colours there is version with suffix _ln which appends a newline (similar to print! and println!).

There are also prnt! and prnt_ln! available which print using the current default foreground colour.

Every variant also has a version with prefix e_. Variants with this prefix output to stderr, while those without output to stdout.

Macros

black

Macro similar to print! using black as the foreground colour.

black_ln

Macro similar to println! using black as the foreground colour.

blue

Macro similar to print! using blue as the foreground colour.

blue_ln

Macro similar to println! using blue as the foreground colour.

cyan

Macro similar to print! using cyan as the foreground colour.

cyan_ln

Macro similar to println! using cyan as the foreground colour.

dark_blue

Macro similar to print! using dark blue as the foreground colour.

dark_blue_ln

Macro similar to println! using dark blue as the foreground colour.

dark_cyan

Macro similar to print! using dark cyan as the foreground colour.

dark_cyan_ln

Macro similar to println! using dark cyan as the foreground colour.

dark_green

Macro similar to print! using dark green as the foreground colour.

dark_green_ln

Macro similar to println! using dark green as the foreground colour.

dark_grey

Macro similar to print! using dark grey as the foreground colour.

dark_grey_ln

Macro similar to println! using dark grey as the foreground colour.

dark_magenta

Macro similar to print! using dark magenta as the foreground colour.

dark_magenta_ln

Macro similar to println! using dark magenta as the foreground colour.

dark_red

Macro similar to print! using dark red as the foreground colour.

dark_red_ln

Macro similar to println! using dark red as the foreground colour.

dark_yellow

Macro similar to print! using dark yellow as the foreground colour.

dark_yellow_ln

Macro similar to println! using dark yellow as the foreground colour.

e_black

Macro similar to eprint! using black as the foreground colour.

e_black_ln

Macro similar to eprintln! using black as the foreground colour.

e_blue

Macro similar to eprint! using blue as the foreground colour.

e_blue_ln

Macro similar to eprintln! using blue as the foreground colour.

e_cyan

Macro similar to eprint! using cyan as the foreground colour.

e_cyan_ln

Macro similar to eprintln! using cyan as the foreground colour.

e_dark_blue

Macro similar to eprint! using dark blue as the foreground colour.

e_dark_blue_ln

Macro similar to eprintln! using dark blue as the foreground colour.

e_dark_cyan

Macro similar to eprint! using dark cyan as the foreground colour.

e_dark_cyan_ln

Macro similar to eprintln! using dark cyan as the foreground colour.

e_dark_green

Macro similar to eprint! using dark green as the foreground colour.

e_dark_green_ln

Macro similar to eprintln! using dark green as the foreground colour.

e_dark_grey

Macro similar to eprint! using dark grey as the foreground colour.

e_dark_grey_ln

Macro similar to eprintln! using dark grey as the foreground colour.

e_dark_magenta

Macro similar to eprint! using dark magenta as the foreground colour.

e_dark_magenta_ln

Macro similar to eprintln! using dark magenta as the foreground colour.

e_dark_red

Macro similar to eprint! using dark red as the foreground colour.

e_dark_red_ln

Macro similar to eprintln! using dark red as the foreground colour.

e_dark_yellow

Macro similar to eprint! using dark yellow as the foreground colour.

e_dark_yellow_ln

Macro similar to eprintln! using dark yellow as the foreground colour.

e_green

Macro similar to eprint! using green as the foreground colour.

e_green_ln

Macro similar to eprintln! using green as the foreground colour.

e_grey

Macro similar to eprint! using grey as the foreground colour.

e_grey_ln

Macro similar to eprintln! using grey as the foreground colour.

e_magenta

Macro similar to eprint! using magenta as the foreground colour.

e_magenta_ln

Macro similar to eprintln! using magenta as the foreground colour.

e_prnt

Macro similar to eprint! using the current default foreground colour.

e_prnt_ln

Macro similar to eprintln! using the current default foreground colour.

e_red

Macro similar to eprint! using red as the foreground colour.

e_red_ln

Macro similar to eprintln! using red as the foreground colour.

e_white

Macro similar to eprint! using white as the foreground colour.

e_white_ln

Macro similar to eprintln! using white as the foreground colour.

e_yellow

Macro similar to eprint! using yellow as the foreground colour.

e_yellow_ln

Macro similar to eprintln! using yellow as the foreground colour.

green

Macro similar to print! using green as the foreground colour.

green_ln

Macro similar to println! using green as the foreground colour.

grey

Macro similar to print! using grey as the foreground colour.

grey_ln

Macro similar to println! using grey as the foreground colour.

magenta

Macro similar to print! using magenta as the foreground colour.

magenta_ln

Macro similar to println! using magenta as the foreground colour.

prnt

Macro similar to print! using the current default foreground colour.

prnt_ln

Macro similar to println! using the current default foreground colour.

red

Macro similar to print! using red as the foreground colour.

red_ln

Macro similar to println! using red as the foreground colour.

white

Macro similar to print! using white as the foreground colour.

white_ln

Macro similar to println! using white as the foreground colour.

yellow

Macro similar to print! using yellow as the foreground colour.

yellow_ln

Macro similar to println! using yellow as the foreground colour.